Navigation fix for pre-scrolled page loads
If the page already has scroll upon page load (in case of page reload or anchor
deep-linking) `getBoundingClientRect` does not return the true value, and needs
to be compensated with the current scroll value.

Also, this will call `menu()` on load, so the navigation menu is open according
to its usual behavior.